requirements.txt
django-environ
Pillow
requests
cryptography